+++ CHANGELOG.md
@@ -10,6 +10,8 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 ### Added
 
 - Support for Erlang, Haskell, R, and SQL programming languages has been added.
+- A new `exclude-dir` flag was added to `todos` that allows for excluding
+  matching directories from the search.
 
 ## [0.4.0] - 2023-08-23